src-py/requirements.txt
alembic==1.13.1
aiosqlite
annotated-types==0.7.0
anyio
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cssselect2==0.8.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
greenlet==3.2.4
h11
html5lib==1.1
httpcore
httptools==0.6.4
httpx==0.23.0
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.2
oscrypto==1.3.0
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.3.4
pydantic_core==2.33.2
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==6.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy[asyncio]
starlette==0.37.2
structlog==24.1.0
svglib==1.5.1
tinycss2==1.4.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
geopy
pytest
pytest-asyncio